requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
opencv-python-headless==4.8.0.76
numpy==1.24.2
faiss-cpu==1.7.3
elasticsearch==8.9.0
SQLAlchemy==2.0.20
